adding no_purge to cloudflare (#110)

+++ providers/cloudflare/cloudflareProvider.go
@@ -103,6 +103,9 @@ func (c *CloudflareApi) GetDomainCorrections(dc *models.DomainConfig) ([]*models
 	corrections := []*models.Correction{}
 
 	for _, d := range del {
+		if dc.KeepUnknown {
+			continue
+		}
 		corrections = append(corrections, c.deleteRec(d.Existing.Original.(*cfRecord), id))
 	}
 	for _, d := range create {
